+++ src/main/java/br/com/CalculcadoraCr/service/CalculoCrService.java
@@ -5,14 +5,13 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CalculoCrService {
-
     public double calcularCr(CalculoCrRequest request) {
         double somaNotas = 0.0;
-        int somaHoras = 0;
+        double somaHoras = 0;
 
         for (int i = 0; i < request.getNotas().size(); i++) {
             double nota = request.getNotas().get(i).getValor();
-            int horas = request.getCargasHorarias().get(i).getHoras();
+            double horas = request.getCargasHorarias().get(i).getHoras();
             somaNotas += nota * horas;
             somaHoras += horas;
         }
